+++ src/main/java/study/stringcalculator/domain/InputValue.java
@@ -1,4 +1,4 @@
-package stringcalculator.domain;
+package study.stringcalculator.domain;
 
 public class InputValue {
