+++ src/test/java/study/stringcalculator/domain/OperatorTest.java
@@ -1,9 +1,9 @@
-package stringcalculator.domain;
-
-import static org.assertj.core.api.Assertions.assertThat;
+package study.stringcalculator.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class OperatorTest {
