+++ src/test/java/study/stringcalculator/factory/OperatorFactoryTest.java
@@ -1,12 +1,12 @@
-package stringcalculator.factory;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+package study.stringcalculator.factory;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import stringcalculator.domain.Operator;
+import study.stringcalculator.domain.Operator;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class OperatorFactoryTest {
