+++ src/test/java/study/stringcalculator/service/CalculateServiceTest.java
@@ -1,32 +1,25 @@
-package stringcalculator.domain;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+package study.stringcalculator.service;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import study.stringcalculator.domain.InputValue;
 
-public class ExpressionTest {
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-  @DisplayName("연산자로 끝나는 연산식의 길이는 짝수이므로 길이가 짝수면 Exception을 던진다.")
-  @ParameterizedTest
-  @ValueSource(strings = {"1 +", "2 - 2 +"})
-  public void expression_ThrowsException_lengthEven(String input) {
-    String[] expession = new InputValue(input).split();
+public class CalculateServiceTest {
 
-    assertThatThrownBy(() -> new Expression(expession))
-        .isInstanceOf(IllegalArgumentException.class);
-  }
+  CalculateService calculateService = new CalculateService();
 
   @DisplayName("숫자가 와야할 순서에 숫자가 아닌 문자를 만나면 IllegalArgumentException을 던진다.")
   @ParameterizedTest
   @ValueSource(strings = {"1 + +", "+ 1 +", "+ + 1"})
   public void calculate_throwException_outOfOrderNumber(String input) {
-    String[] expession = new InputValue(input).split();
+    InputValue inputValue = new InputValue(input);
 
-    assertThatThrownBy(() -> new Expression(expession).calculate())
+    assertThatThrownBy(() -> calculateService.calculate(inputValue))
         .isInstanceOf(IllegalArgumentException.class);
   }
 
@@ -34,9 +27,19 @@ public class ExpressionTest {
   @ParameterizedTest
   @ValueSource(strings = {"1 1 1", "1 ? 1", "1 ^ 1"})
   public void calculate_throwException_outOfOrderOperator(String input) {
-    String[] expession = new InputValue(input).split();
+    InputValue inputValue = new InputValue(input);
 
-    assertThatThrownBy(() -> new Expression(expession).calculate())
+    assertThatThrownBy(() -> calculateService.calculate(inputValue))
+        .isInstanceOf(IllegalArgumentException.class);
+  }
+
+  @DisplayName("사칙연산으로 끝나는 식은 IllegalArgumentException을 던진다.")
+  @ParameterizedTest
+  @ValueSource(strings = {"1 +", "1 -", "1 *"})
+  public void calculate_throwException_EndOperator(String input) {
+    InputValue inputValue = new InputValue(input);
+
+    assertThatThrownBy(() -> calculateService.calculate(inputValue))
         .isInstanceOf(IllegalArgumentException.class);
   }
 
@@ -44,8 +47,8 @@ public class ExpressionTest {
   @ParameterizedTest
   @CsvSource({"1 + 1,2", "2 - 1,1", "2 * 2,4", "4 / 3,1", "3 / 5 + 100,100", "1 + 1 + 1,3"})
   public void calculate_Result(String input, int result) {
-    String[] expession = new InputValue(input).split();
+    InputValue inputValue = new InputValue(input);
 
-    assertThat(new Expression(expession).calculate()).isEqualTo(result);
+    assertThat(calculateService.calculate(inputValue)).isEqualTo(result);
   }
 }
